pyproject.toml
[project]
name = "lox"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lark-parser>=0.12.0",
    "pytest>=8.3.5",
    "pytest-json-report>=1.5.0",
    "rich>=14.0.0",
]

[project.scripts]
lox = "lox.cli:main"

[dependency-groups]
dev = ["pytest>=8.3.5", "pytest-timeout>=2.4.0"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["F403", "F405"]
